Running ng1 and ng2 at the same time
