+++ orcid-web/src/main/webapp/static/javascript/ng1Orcid/app/modules/ng1_app.ts
@@ -6,3 +6,17 @@
 //var orcidNgModule = angular.module('orcidApp', ['ngCookies','ngSanitize', 'ui.multiselect', 'vcRecaptcha']);
 
 //angular.bootstrap(document.body, ['orcidApp']);
+
+/**
+ * This file defines the root module of the Angular 1 of the application.
+ */
+//import 'reflect-metadata';
+// import angular1
+import * as angular from 'angular'
+import 'angular-route'
+
+// import app modules
+//import {MessagesModule} from './messages';
+//import {MenuModule} from './menu';
+
+export const Ng1AppModule = angular.module('Ng1AppModule', ['ngRoute']);
