+++ orcid-web/src/main/webapp/static/javascript/ng1Orcid/app/modules/ng2_app.ts
@@ -1,0 +1,26 @@
+import 'reflect-metadata';
+
+import {NgModule, Component} from '@angular/core';
+import {BrowserModule} from '@angular/platform-browser';
+import {UpgradeModule} from '@angular/upgrade/static';
+/*
+@Component({
+  selector: 'root-cmp',
+  template: `
+    <div class="ng-view"></div>
+  `,
+}) 
+export class RootCmp {}
+*/
+@NgModule({
+  imports: [
+    BrowserModule,
+    UpgradeModule
+  ]//,
+  //bootstrap: [RootCmp],
+  //declarations: [RootCmp]
+})
+export class Ng2AppModule {
+  //constructor(){}
+  ngDoBootstrap() {}
+}
